fix: renamed and moved carousel components
