+++ src/Components/Carousel/Arrow/Arrow.css
@@ -1,6 +1,6 @@
 .arrowIcon-left,
 .arrowIcon-right {
-	content: url(../../utils/img/arrow-icon.png);
+	content: url(../../../utils/img/arrow-icon.png);
 }
 
 .arrowIcon-left {
