+++ src/Components/Carousel/Bullets/Bullet/Bullet.css
