--- src/Components/CarouselBullets/CarouselBullets.js
@@ -1,20 +1,0 @@
-import React from "react";
-import "./CarouselBullets.css";
-
-import CarouselBullet from "../CarouselBullet/CarouselBullet";
-
-function CarouselBullets(props) {
-  return (
-    <div className="bullets-wrapper--carousel">
-      {props.arrOfImages.map((element, key) => {
-        return key === props.active ? (
-          <CarouselBullet active={true} key={key} index={key} />
-        ) : (
-          <CarouselBullet active={false} key={key} index={key} />
-        );
-      })}
-    </div>
-  );
-}
-
-export default CarouselBullets;
